03_Tools/pyproject.toml
[project]
name = "03-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.2",
    "langchain-openai>=1.0.1",
    "python-dotenv>=1.1.1",
]
